+++ server/motd.go
